+++ flutter/lib/authentication/authentication.dart
@@ -1,0 +1,4 @@
+export 'blocs/user_bloc.dart';
+export 'models/user.dart';
+export 'persistence/user_preferences_repository.dart';
+export 'repositories/user_repository.dart';
