+++ flutter/lib/authentication/models/user.dart
