+++ flutter/lib/authentication/repositories/user_repository.dart
@@ -1,0 +1,9 @@
+import 'package:journey/authentication/authentication.dart';
+
+abstract class UserRepository {
+  Future<User?> getUser();
+
+  Future<void> removeUser();
+
+  Future<void> setUser(User entry);
+}
