+++ flutter/lib/dashboard/dashboard.dart
@@ -1,0 +1,1 @@
+export './view/dashboard_screen.dart';
