+++ flutter/lib/entries/entries.dart
@@ -1,0 +1,6 @@
+export 'blocs/entries_bloc.dart';
+export 'models/entry.dart';
+export 'repositories/entry_repository.dart';
+export 'persistence/entry_database_repository.dart';
+export 'view/add_entry_screen.dart';
+export 'view/entries_list.dart';
