+++ flutter/lib/entries/repositories/entry_repository.dart
@@ -1,0 +1,11 @@
+import 'package:journey/entries/models/entry.dart';
+
+abstract class EntryRepository {
+  Future<Entry> insert(Entry entry);
+
+  Future<Entry> update(Entry entry);
+
+  Future<Entry> delete(Entry entry);
+
+  Future<List<Entry>> getEntries();
+}
