+++ flutter/lib/entries/view/add_entry_screen.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+import './add_entry_form/add_entry_form.dart';
+
+class AddEntryScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Add Entry"),
+      ),
+      body: AddEntryForm(),
+    );
+  }
+}
