+++ flutter/lib/entries/view/entry_item.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+
+import '../models/entry.dart';
+
+class EntryItem extends StatelessWidget {
+  final Entry entry;
+  final GestureTapCallback onTap;
+
+  EntryItem({required this.entry, required this.onTap})
+      : super(key: Key(entry.id));
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+        onTap: onTap,
+        child: Column(children: [
+          Text(entry.title, key: Key('${entry.id}__title')),
+          Text(entry.location.formattedLocation(),
+              key: Key('${entry.id}__subtitle'))
+        ]));
+    // return Scaffold(
+    //     body: ListTile(
+    //   onTap: onTap,
+    //   title: Text(entry.title, key: Key('${entry.id}__title')),
+    //   subtitle: Text(entry.location.formattedLocation(),
+    //       key: Key('${entry.id}__subtitle')),
+    // ));
+  }
+}
