--- flutter/lib/repositories/entry_repository.dart
@@ -1,11 +1,0 @@
-import 'package:journey/models/entry.dart';
-
-abstract class EntryRepository {
-  Future<Entry> insert(Entry entry);
-
-  Future<Entry> update(Entry entry);
-
-  Future<Entry> delete(Entry entry);
-
-  Future<List<Entry>> getEntries();
-}
