--- flutter/lib/repositories/user_repository.dart
@@ -1,9 +1,0 @@
-import 'package:journey/models/user.dart';
-
-abstract class UserRepository {
-  Future<User?> getUser();
-
-  Future<void> removeUser();
-
-  Future<void> setUser(User entry);
-}
