--- flutter/lib/screens/add_entry_screen.dart
@@ -1,14 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:journey/widgets/add_entry_form/add_entry_form.dart';
-
-class AddEntryScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text("Add Entry"),
-      ),
-      body: AddEntryForm(),
-    );
-  }
-}
