+++ flutter/lib/splash/splash.dart
@@ -1,0 +1,1 @@
+export 'view/splash_screen.dart';
