+++ flutter/pubspec.yaml
@@ -29,7 +29,7 @@ dependencies:
     cupertino_icons: ^1.0.3
     english_words: ^4.0.0
     equatable: ^2.0.0
-    flutter_bloc: ^7.0.0
+    flutter_bloc: ^7.3.0
     shared_preferences: ^2.0.8
     sqflite: ^2.0.0+3
